Delete unnecessary SQL files
asterisk_db.sql and asterisk_role.sql where created to easily create
the database and the database-role. This is no longer necessary as it
is implemented in the script located in /scripts/modules/postgresql_setup.sh.
The script is run on install of Asterisk and sets up the database accordingly.