Updated README files
The purpose of this repository started out as a devcontainer and is now
that, as well as an Ubuntu install of Asterisk. To clear up some of this
I have written a more general README.md and split the other files
into DEVCONTAINER.md and UBUNTU.md.

For now, the files README-OLD.md and POSTGRES_ASTERISK.md
still exist, and they look confusing. They will be removed and their
contents put into DEVCONTAINER.md and UBUNTU.md.